pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "diva"
version = "0.1.0"
description = "DIVA — Enterprise Multi-Agent Chat System"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-ollama>=0.2.0",
    "langgraph>=0.2.0",
    "mcp>=1.0.0",
    "motor>=3.6.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "sse-starlette>=2.0.0",
    "python-dotenv>=1.0.0",
    "cachetools>=5.5.0",
    "structlog>=24.0.0",
    "neo4j>=5.0.0",
    "certifi>=2024.0.0",
]

[project.optional-dependencies]
eval = [
    "deepeval>=1.3.0",
    "ragas>=0.2.0",
]
tachyon = [
    "langchain-openai>=0.2.0",
]
mcp-servers = [
    "fastmcp>=3.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/diva", "src/tachyon_langchain_client"]
